Clean up versioned package files
